Case study and project management
Strangely, the final project of my ICT Project Management unit at University was to plan and manage a wedding. Whilst it seemed strange that an Information Technology bachelor degree would task me with planning a wedding, the principles and management methodologies are the same as those required to plan and manage an ICT project.